refactor: merge setMaintenance code with similar in other containers
